* Modify memory map to be compatible with Disk Basic
Change memory map so the game code is loaded at $E00 and the data area
starts at $4200 and runs to $7FFF. This allows the usual Disk Basic stuff in
low memory to operate without interference. It also allows the code to grow
upward to about 13K in size without having to relocate the data, which is
painful since the graphic screens MUST be on 512 byte boundaries.

Also added some symbolic definitions to make changing the graphic screen
locations less problematic.