Added autostart to the BIN file
Created an autostart routine that stops the disk drive and starts the game
automatically. It intercepts RVEC12 (line input) to start the game. Since
neither version of Disk Basic actually does anything with that vector (it's
an RTS) and Daggorath completely takes over anyway, there's no need to
preserve it. (Even the bits of the ROMs that Daggorath uses don't rely on
the line input routine.)