Changed ZSAVE to save to disk.
Added a save routine for saving games to disk. This completely replaces the
original save routine. Unlike the original save, it does not just to a
binary dump of the game state. Instead, it makes an attempt to keep the
state generic enough that it can be reloaded by a game with different
variables.