Remove unused tape code and reset PIAs to Basic mode before load/save
Remove unused tape code now that saving to disk is supported.

Also, change the default PIA setting to make the RS232 output bit to an
input. This should prevent junk on the serial port duing the game and may
improve the situation with drivewire on the bitbanger versions of Disk
Basic.

Additional changes to the routines used to set up for calls to and returns
from Disk Basic ROMs were required to re-enable the RS232 output. Also, to
prevent weird sounds during Drivewire bitbanger operations, set the single
bit sound to an input during save/load and reset to output afterward.