Remove SWI2 ROM call vector
With the removal of the tape code, there is no more point in having the SWI2
ROM call handler. With only a single call through the Color Basic jump
table, it is simpler to just replace the SWI2 call sequence with a sequence
to set and restore DP for Basic and just call the indirect vector directly.

This also removes the unreachable remnant of the same code from among the
SWI calls.