Squash of four back-ported commits:
* Make graphics data more readable with some labels and macros

Created macros for graphics data to make the meaning clearer. Also added
descriptive labels to most graphics data.

* Fix missing vertex on open right passage image

* Change SWI routines to just proper jump table

Change the routine vector table for SWI calls to use a proper jump table
instead of a sequence of offsets from the previous routine. This allows the
routines to be reorganized and makes access to the routines run faster. This
does have an impact on game play because less time is taken looking up
routine addresses.

* Fix missing vertex on "door to the right" graphic