Squashed two commits:
* Convert font to unpacked representation

Convert the text font to an unpacked raw format. Also include glyphs for all
the regular ASCII characters in preparation for changing over to full ASCII
operations.

* Include missing file and Makefile modifications

Need to actually include the font data if it's going to be referenced. Also
update dependencies in dod.s to reflect the additional file.