Fix bug saving games.
There was a missing table entry in the scheduling table entry fixup list.
This caused save games to break when the data structures in the direct page
and above were shifted around. The missing entry has been added.

An additional entry was added to allow loading previously saved games where
the fixup table was wrong.

Finally, move the pause mode variables back into the direct page as it is no
longer required that they be elsewhere.